backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
alembic==1.15.1
pydantic[email]==2.11.1
python-jose[cryptography]==3.4.0
bcrypt==4.3.0
python-multipart==0.0.20
pdfplumber==0.11.6
python-docx==1.1.2
pandas==2.2.3
numpy==2.2.4
scikit-learn==1.8.0
spacy==3.8.4
google-generativeai==0.8.4
python-dotenv==1.0.1
psycopg[binary]==3.2.10
boto3==1.39.13
